refactor: split tabs.go to stay under the file-length cap (TASK-879)
tabs.go sat at exactly the 800-line revive cap, leaving no room to add
the pending-fetch marker TASK-879 needs. Move the tab lifecycle
functions (saveCurrentTab, loadTab, cloneCurrentTab, moveActiveTab,
tab UID handling, tab labels) into a new tab_state.go; tabs.go keeps
the smaller navigation/status-message/error-log helpers.

Mechanical extraction only, no behavior change.
